address comments

+++ src/token/IToken.sol
@@ -82,7 +82,7 @@ interface IToken is IUUPS, IERC721Votes, TokenTypesV1, TokenTypesV2 {
     function mint(address recipient) external returns (uint256 tokenId);
 
     /// @notice Mints the specified amount of tokens to the recipient and handles founder vesting
-    function mint(uint16 amount, address recipient) external returns (uint256[] memory tokenIds);
+    function mint(uint256 amount, address recipient) external returns (uint256[] memory tokenIds);
 
     /// @notice Burns a token owned by the caller
     /// @param tokenId The ERC-721 token id
